# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.CalculatorRequest:
    properties:
      expression:
        description: 计算表达式
        type: string
    required:
    - expression
    type: object
  controllers.CalculatorResponse:
    properties:
      error:
        type: string
      result:
        type: string
    type: object
...
  controllers.LoginDTO:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  controllers.RegisterDTO:
    properties:
      password:
        maxLength: 64
        minLength: 6
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  controllers.deleteInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: Go-Web综合性应用API接口文档
  title: Go-Web项目 API
  version: 0.0.3
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: token,result_url
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户登录
      tags:
      - Auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 退出登录
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户注册
      tags:
      - Auth
  /calculator/calculate:
    post:
      consumes:
      - application/json
      description: 支持基本的四则运算，包括小数点计算
      parameters:
      - description: 计算表达式
        in: body
        name: expression
        required: true
        schema:
          $ref: '#/definitions/controllers.CalculatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.CalculatorResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 计算器计算
      tags:
      - Calculator
...
  /user/delete:
    delete:
      consumes:
      - application/json
      description: Delete the current user's account after password verification
      parameters:
      - description: User credentials
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.deleteInput'
      produces:
      - application/json
      responses:
        "200":
          description: 退出成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - User
...
swagger: "2.0"